Split `core` and `lib` in `include/kithare`
