+++ include/kithare/core/ast.h
@@ -13,8 +13,8 @@ extern "C" {
 #include <stdbool.h>
 #include <stdint.h>
 
-#include "array.h"
-#include "string.h"
+#include <kithare/lib/array.h>
+#include <kithare/lib/string.h>
 
 
 typedef struct _khAst khAst;
