+++ include/kithare/core/token.h
@@ -12,8 +12,8 @@ extern "C" {
 
 #include <stdint.h>
 
-#include "array.h"
-#include "string.h"
+#include <kithare/lib/array.h>
+#include <kithare/lib/string.h>
 
 
 typedef enum {
