+++ include/kithare/lib/array.h
@@ -23,15 +23,23 @@ typedef struct {
 } _khArrayHeader;
 
 
-#define khArray(TYPE) TYPE*
+// Getter macros from the header
+#define khArray(TYPE) TYPE* // Type alias
 #define khArray_header(ARRAY) (((_khArrayHeader*)(*ARRAY))[-1])
 #define khArray_typeSize(ARRAY) (khArray_header(ARRAY).type_size)
 #define khArray_size(ARRAY) (khArray_header(ARRAY).size)
 #define khArray_reserved(ARRAY) (khArray_header(ARRAY).reserved)
 #define khArray_deleter(ARRAY) (khArray_header(ARRAY).deleter)
 
-void* _khArray_new(size_t type_size, void (*deleter)(void*));
+
 #define khArray_new(TYPE, DELETER) (TYPE*)_khArray_new(sizeof(TYPE), (void (*)(void*))(DELETER))
+static inline void* _khArray_new(size_t type_size, void (*deleter)(void*)) {
+    // Don't forget to allocate an extra null-terminator space in case it's a string
+    void* array = calloc(sizeof(_khArrayHeader) + type_size, 1);
+    *(_khArrayHeader*)array =
+        (_khArrayHeader){.type_size = type_size, .size = 0, .reserved = 0, .deleter = deleter};
+    return array + sizeof(_khArrayHeader);
+}
 
 // The gnu11 standard must be used with some of the macros here. Since some utilize `typeof` and
 // statement-expressions, which are not available in vanilla C
@@ -66,6 +74,7 @@ void* _khArray_new(size_t type_size, void (*deleter)(void*));
         __kh_copy;                                                                                 \
     })
 
+
 // Verifies that the argument given is a pointer to a pointer (e.g: int**, char**), then casts it into
 // a void**
 #define _kh_verifyPtrToPtr(PTR)                                            \
@@ -75,21 +84,94 @@ void* _khArray_new(size_t type_size, void (*deleter)(void*));
         (void**)__kh_ptr;                                                  \
     })
 
-void _khArray_delete(void** array);
+
 #define khArray_delete(ARRAY) _khArray_delete(_kh_verifyPtrToPtr(ARRAY))
 #define khArray_arrayDeleter(TYPE) ((void (*)(TYPE**))_khArray_delete)
+static inline void _khArray_delete(void** array) {
+    if (khArray_deleter(array) != NULL) {
+        // Increment by type size
+        for (size_t i = 0; i < khArray_typeSize(array) * khArray_size(array);
+             i += khArray_typeSize(array)) {
+            khArray_deleter(array)(*array + i);
+        }
+    }
 
-void _khArray_reserve(void** array, size_t size);
+    // Don't forget to undo the header offset before freeing it
+    free(*array - sizeof(_khArrayHeader));
+    *array = NULL;
+}
+
 #define khArray_reserve(ARRAY, SIZE) _khArray_reserve(_kh_verifyPtrToPtr(ARRAY), SIZE)
+static inline void _khArray_reserve(void** array, size_t size) {
+    if (khArray_reserved(array) >= size) {
+        return;
+    }
 
-void _khArray_fit(void** array);
+    // Also, don't forget the null-terminator space
+    void* expanded_array = calloc(sizeof(_khArrayHeader) + khArray_typeSize(array) * (size + 1), 1);
+    *(_khArrayHeader*)expanded_array = khArray_header(array);
+    memcpy(expanded_array + sizeof(_khArrayHeader), *array,
+           khArray_typeSize(array) * khArray_size(array));
+    // Undo offset, free!
+    free(*array - sizeof(_khArrayHeader));
+
+    ((_khArrayHeader*)expanded_array)->reserved = size;
+    *array = expanded_array + sizeof(_khArrayHeader);
+}
+
 #define khArray_fit(ARRAY) _khArray_fit(_kh_verifyPtrToPtr(ARRAY));
+static inline void _khArray_fit(void** array) {
+    // Pretty much the same implementation of `_khArray_reserve` but with this part different
+    if (khArray_reserved(array) == khArray_size(array)) {
+        return;
+    }
 
-void _khArray_pop(void** array, size_t items);
+    void* shrunk_array =
+        calloc(sizeof(_khArrayHeader) + khArray_typeSize(array) * (khArray_size(array) + 1), 1);
+    *(_khArrayHeader*)shrunk_array = khArray_header(array);
+    memcpy(shrunk_array + sizeof(_khArrayHeader), *array,
+           khArray_typeSize(array) * khArray_size(array));
+    free(*array - sizeof(_khArrayHeader));
+
+    ((_khArrayHeader*)shrunk_array)->reserved = khArray_size(array);
+    *array = shrunk_array + sizeof(_khArrayHeader);
+}
+
 #define khArray_pop(ARRAY, ITEMS) _khArray_pop(_kh_verifyPtrToPtr(ARRAY), ITEMS)
+static inline void _khArray_pop(void** array, size_t items) {
+    // If it's trying to pop more items than the array actually has, cap it
+    items = items > khArray_size(array) ? khArray_size(array) : items;
 
-void _khArray_reverse(void** array);
+    if (khArray_deleter(array) != NULL) {
+        for (size_t i = items; i > 0; i--) {
+            khArray_deleter(array)(*array + khArray_typeSize(array) * (khArray_size(array) - i));
+        }
+    }
+
+    // Clean after yourself
+    memset(*array + khArray_typeSize(array) * (khArray_size(array) - items), 0,
+           khArray_typeSize(array) * items);
+    khArray_size(array) -= items;
+}
+
 #define khArray_reverse(ARRAY) _khArray_reverse(_kh_verifyPtrToPtr(ARRAY))
+static inline void _khArray_reverse(void** array) {
+    for (size_t i = 0; i < khArray_size(array) / 2; i++) {
+        char tmp[khArray_typeSize(array)];
+
+        // front -> tmp
+        memcpy(tmp, *array + i * khArray_typeSize(array), khArray_typeSize(array));
+
+        // back -> front
+        memcpy(*array + i * khArray_typeSize(array),
+               *array + (khArray_size(array) - i - 1) * khArray_typeSize(array),
+               khArray_typeSize(array));
+
+        // tmp -> back
+        memcpy(*array + (khArray_size(array) - i - 1) * khArray_typeSize(array), tmp,
+               khArray_typeSize(array));
+    }
+}
 
 #define khArray_memory(ARRAY, PTR, SIZE, COPIER)                                                      \
     {                                                                                                 \
