+++ include/kithare/lib/info.h
